Revert "fix: fixed data type classes and library core files"
This reverts commit 493398be1933f3b4e2346adeaacfe22040c243aa.

+++ route4me-csharp-sdk/Route4MeSDKLibrary/DataTypes/DeviceLocationTracePoint.cs
@@ -10,7 +10,7 @@ namespace Route4MeSDK.DataTypes
 
         [DataMember(Name = "waypoint_index")] public int? WaypointIndex { get; set; }
 
-        [DataMember(Name = "location")] public double[] Location { get; set; }
+        [DataMember(Name = "location")] public int[] Location { get; set; }
 
         [DataMember(Name = "name")] public string Name { get; set; }
